fix flightlistpage component & install axios for fetch
