+++ frontend-app-skynuc/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.20.0",
     "classnames": "^2.2.6",
     "html-to-react-components": "^1.6.6",
     "moment": "^2.27.0",
